adding prob limit and not_in_lexicon clauses
