+++ djAerolith/lib/query_generator/exceptions.py
@@ -1,0 +1,2 @@
+class BadInput(Exception):
+    pass
