+++ djAerolith/lib/word_searches.py
@@ -3,6 +3,9 @@ Helper functions for dealing with word searches.
 
 """
 import logging
+
+from base.models import Lexicon
+
 logger = logging.getLogger(__name__)
 
 
@@ -18,10 +21,11 @@ class SearchDescription(object):
     MATCHING_ANAGRAM = 'matching_anagram'
     ALPHAGRAM_LIST = 'alphagram_list'
     LEXICON = 'lexicon'
+    NOT_IN_LEXICON = 'not_in_lexicon'
 
     # Must always be the first search description.
     @staticmethod
-    def lexicon(lex):
+    def lexicon(lex: Lexicon):
         return {
             "condition": SearchDescription.LEXICON,
             "lexicon": lex,
@@ -116,8 +120,15 @@ class SearchDescription(object):
             "letters": letters
         }
 
+    @staticmethod
+    def not_in_lexicon(lex: str):
+        return {
+            'condition': SearchDescription.NOT_IN_LEXICON,
+            'lexicon': lex,
+        }
 
-def temporary_list_name(search_descriptions):
+
+def temporary_list_name(search_descriptions, lexicon_name):
     """ Build up a temporary list name given a search description. """
     tokens = []
     for sd in search_descriptions:
@@ -125,34 +136,46 @@ def temporary_list_name(search_descriptions):
             tokens.append(sd['lexicon'].lexiconName)
         elif sd['condition'] == SearchDescription.LENGTH:
             if sd['min'] == sd['max']:
-                tokens.append('{}s'.format(sd['min']))
+                tokens.append(f'{sd["min"]}s')
             elif sd['min'] < sd['max']:
                 tk = []
                 for i in range(sd['min'], sd['max'] + 1):
-                    tk.append('{}s'.format(i))
+                    tk.append(f'{i}s')
                 tokens.append(', '.join(tk))
             else:
                 tokens.append('INVALID')
         elif sd['condition'] == SearchDescription.PROB_RANGE:
-            tokens.append('({} - {})'.format(sd['min'], sd['max']))
+            tokens.append(f'({sd["min"]} - {sd["max"]})')
         elif sd['condition'] == SearchDescription.NUM_ANAGRAMS:
             if sd['min'] == 1 and sd['max'] == 1:
                 tokens.append('Single-anagram')
             else:
-                tokens.append('#-anagrams: {} - {}'.format(sd['min'],
-                                                           sd['max']))
+                tokens.append(f'#-anagrams: {sd["min"]} - {sd["max"]}')
         elif sd['condition'] == SearchDescription.POINT_VALUE:
             if sd['min'] == sd['max']:
-                tokens.append('{}-pt'.format(sd['min']))
+                tokens.append(f'{sd["min"]}-pt')
             else:
-                tokens.append('pts: {} - {}'.format(sd['min'], sd['max']))
+                tokens.append(f'pts: {sd["min"]} - {sd["max"]}')
         elif sd['condition'] == SearchDescription.NUM_VOWELS:
             if sd['min'] == sd['max']:
-                tokens.append('{}-vowel'.format(sd['min']))
+                tokens.append(f'{sd["min"]}-vowel')
             else:
-                tokens.append('vowels: {} - {}'.format(sd['min'], sd['max']))
+                tokens.append(f'vowels: {sd["min"]} - {sd["max"]}')
         elif sd['condition'] == SearchDescription.HAS_TAGS:
-            tokens.append('tags: {}'.format(', '.join(sd['tags'])))
+            tokens.append(f'tags: {", ".join(sd["tags"])}')
         elif sd['condition'] == SearchDescription.PROB_LIMIT:
             tokens.append(f'(limit {sd["min"]} - {sd["max"]})')
+        elif sd['condition'] == SearchDescription.NOT_IN_LEXICON:
+            desc = ''
+            if sd['lexicon'] == 'other_english':
+                if lexicon_name == 'America':
+                    desc = 'CSW15'
+                elif lexicon_name == 'CSW15':
+                    desc = 'America'
+            elif sd['lexicon'] == 'update':
+                if lexicon_name == 'America':
+                    desc = 'OWL2'
+                elif lexicon_name == 'CSW15':
+                    desc = 'CSW12'
+            tokens.append(f'not in {desc}')
     return ' '.join(tokens)
